Add new API to enumerate available SocketCAN interfaces
This patch adds a new API, `enumerate_interfaces()`, which can be used
to obtain a list of available SocketCAN interfaces. It is available
under the `enumerate` feature flag since it pulls the libudev dependency
(it may not be available in all Linux systems).

This feature is inspired by the `enumerate_ports()` API found in
serialport-rs.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>